+++ web/lib/modalities/music-gen/music-gen.ts
@@ -1,3 +1,5 @@
+import { fetchAPI } from "@/lib/pulse-editor-website/backend";
+
 export class BaseMusicGen {
   private model: any;
   private generateFunc?: (
@@ -95,8 +97,7 @@ export function getMusicGenModel(
         ? (modelName as `${string}/${string}:${string}`)
         : (modelName as `${string}/${string}`);
 
-    const proxyHost = process.env.NEXT_PUBLIC_BACKEND_URL;
-    const response = await fetch(`${proxyHost}/api/inference/replicate`, {
+    const response = await fetchAPI(`/api/inference/replicate`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json",
@@ -122,8 +123,8 @@ export function getMusicGenModel(
       prediction.status !== "failed"
     ) {
       await new Promise((resolve) => setTimeout(resolve, 1000));
-      const response = await fetch(
-        `${proxyHost}/api/inference/replicate/${prediction.id}`,
+      const response = await fetchAPI(
+        `/api/inference/replicate/${prediction.id}`,
         {
           method: "POST",
           body: JSON.stringify({