+++ npm-packages/react-api/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pulse-editor/react-api
 
+## 0.1.1-alpha.26
+
+### Patch Changes
+
+- Fix export alias
+- Updated dependencies
+  - @pulse-editor/shared-utils@0.1.1-alpha.26
+
 ## 0.1.1-alpha.25
 
 ### Patch Changes